// apps/main/src/common/infrastructure/database/types/db.d.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface WsSchoolEstimationDetails {
  available_vaccinator: Generated<number | null>;
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  id: Generated<number>;
  microplanning_id: number;
  notes: string | null;
  required_service: Generated<number | null>;
  required_service_days: Generated<number | null>;
  schedule_month: string | null;
  school_id: number | null;
  status: Generated<number | null>;
  updated_at: Generated<Date>;
  updated_by: number | null;
}
// ...
export interface DB {
// ...
  ws_school_estimation_details: WsSchoolEstimationDetails;
// ...
}